[WIP] Don't create a separate SyncWorker for every sync (run directly within onetime/periodic sync instead)
