+++ app/src/main/kotlin/at/bitfire/davdroid/syncadapter/PeriodicSyncWorker.kt
@@ -15,9 +15,7 @@ import androidx.work.NetworkType
 import androidx.work.Operation
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
-import androidx.work.Worker
 import androidx.work.WorkerParameters
-import at.bitfire.davdroid.InvalidAccountException
 import at.bitfire.davdroid.log.Logger
 import at.bitfire.davdroid.settings.AccountSettings
 import dagger.assisted.Assisted
@@ -28,13 +26,13 @@ import java.util.concurrent.TimeUnit
 /**
  * Handles scheduled sync requests.
  *
- * Enqueues immediate [SyncWorker] syncs at the appropriate moment.
- *
  * The different periodic sync workers each carry a unique work name composed of the account and
  * authority which they are responsible for. For each account there will be multiple dedicated periodic
  * sync workers for each authority. See [PeriodicSyncWorker.workerName] for more information.
  *
  * Deferrable: yes (PeriodicWorkRequest)
+ *
+ * Expedited: no
  *
  * Long-running: no
  */
@@ -42,13 +40,9 @@ import java.util.concurrent.TimeUnit
 class PeriodicSyncWorker @AssistedInject constructor(
     @Assisted appContext: Context,
     @Assisted workerParams: WorkerParameters
-) : Worker(appContext, workerParams) {
+) : BaseSyncWorker(appContext, workerParams) {
 
     companion object {
-        // Worker input parameters
-        internal const val ARG_ACCOUNT_NAME = "accountName"
-        internal const val ARG_ACCOUNT_TYPE = "accountType"
-        internal const val ARG_AUTHORITY = "authority"
 
         /**
          * Unique work name of this worker. Can also be used as tag.
@@ -57,7 +51,7 @@ class PeriodicSyncWorker @AssistedInject constructor(
          *
          * @param account the account this worker is running for
          * @param authority the authority this worker is running for
-         * @return Name of this worker composed as "sync $authority ${account.type}/${account.name}"
+         * @return Name of this worker composed as "periodic-sync $authority ${account.type}/${account.name}"
          */
         fun workerName(account: Account, authority: String): String =
             "periodic-sync $authority ${account.type}/${account.name}"
@@ -85,6 +79,7 @@ class PeriodicSyncWorker @AssistedInject constructor(
                 ).build()
             val workRequest = PeriodicWorkRequestBuilder<PeriodicSyncWorker>(interval, TimeUnit.SECONDS)
                 .addTag(workerName(account, authority))
+                .addTag(commonTag(account, authority))
                 .setInputData(arguments)
                 .setConstraints(constraints)
                 .build()
@@ -110,29 +105,19 @@ class PeriodicSyncWorker @AssistedInject constructor(
 
     }
 
-    override fun doWork(): Result {
-        val account = Account(
-            inputData.getString(ARG_ACCOUNT_NAME) ?: throw IllegalArgumentException("$ARG_ACCOUNT_NAME required"),
-            inputData.getString(ARG_ACCOUNT_TYPE) ?: throw IllegalArgumentException("$ARG_ACCOUNT_TYPE required")
-        )
-        val authority = inputData.getString(ARG_AUTHORITY) ?: throw IllegalArgumentException("$ARG_AUTHORITY required")
-        Logger.log.info("Running periodic sync worker: account=$account, authority=$authority")
+    override suspend fun doSyncWork(
+        account: Account,
+        authority: String,
+        accountSettings: AccountSettings
+    ): Result {
+        Logger.log.info("Running periodic sync for account=$account, authority=$authority")
 
-        val accountSettings = try {
-            AccountSettings(applicationContext, account)
-        } catch (e: InvalidAccountException) {
-            Logger.log.warning("Account $account doesn't exist anymore, cancelling periodic sync")
-            disable(applicationContext, account, authority)
-            return Result.failure()
-        }
-        if (!SyncWorker.wifiConditionsMet(applicationContext, accountSettings)) {
-            Logger.log.info("Sync conditions not met. Won't run sync.")
+        if (!wifiConditionsMet(applicationContext, accountSettings)) {
+            Logger.log.info("WiFi conditions not met. Won't run periodic sync.")
             return Result.failure()
         }
 
-        // Just request immediate sync
-        Logger.log.info("Requesting immediate sync")
-        SyncWorker.enqueue(applicationContext, account, authority, expedited = false)
-        return Result.success()
+        return super.doSyncWork(account, authority, accountSettings)
     }
+
 }
